requirements.txt
python-telegram-bot==20.7
python-dotenv
requests
beautifulsoup4
streamlit
pyyaml
chromadb
pypdf
pymupdf
sentence-transformers
numpy<2
psutil
fastapi
uvicorn
python-multipart
pytest
python-docx
